Switch `qlview` over to a full macOS app.
This stops building `qlview` as a command-line application and makes
it a normal app instead. As an app, it now supports a new URL scheme
`qlview:/path/to/file` that allows opening documents from the
command-line simply via `open`. The advantage of that approach is that
we no longer end up start up separate `qlview` instances for each
document being opened.

We also switch from distributing a ZIP file to a more standard DMG.